+++ app/controllers/EnrollmentController.php
@@ -545,9 +545,9 @@ class EnrollmentController extends Controller
                             $finalMedia = array_sum($arr["semesterMedias"]) / count($arr["semesterMedias"]);
                             if ($grade["unityGrade"] != "") {
                                 $finalRecoverMedia = ($finalMedia + $grade["unityGrade"]) / 2;
-                                $finalMedia = number_format($finalRecoverMedia, 2);
+                                $finalMedia = number_format($finalRecoverMedia, 1);
                             } else {
-                                $finalMedia = number_format($finalMedia, 2);
+                                $finalMedia = number_format($finalMedia, 1);
                                 $rfFilled = false;
                             }
 
@@ -560,7 +560,7 @@ class EnrollmentController extends Controller
                         $media = $sums / $sumsCount;
                         array_push($arr["semesterMedias"], $media);
                     }
-                    $finalMedia = number_format(array_sum($arr["semesterMedias"]) / count($arr["semesterMedias"]), 2);
+                    $finalMedia = number_format(array_sum($arr["semesterMedias"]) / count($arr["semesterMedias"]), 1);
                 }
 
                 //traz a situação do aluno (se null, aprovado, recuperação ou reprovado)
@@ -568,26 +568,26 @@ class EnrollmentController extends Controller
                 $situation = null;
                 if ($allNormalUnitiesFilled) {
                     if ($finalMedia >= $gradeRules->approvation_media) {
-                        $situation = "Aprovado(a)";
+                        $situation = "Aprovado";
                     } else {
                         if ($hasRF) {
                             if ($rfFilled) {
                                 if ($finalMedia >= $gradeRules->final_recover_media) {
-                                    $situation = "Aprovado(a)";
+                                    $situation = "Aprovado";
                                 } else {
-                                    $situation = "Reprovado(a)";
+                                    $situation = "Reprovado";
                                 }
                             } else {
                                 $situation = "Recuperação";
                             }
                         } else if ($recSemIndex > 0) {
                             if ($lastRSFilledGrade !== "") {
-                                $situation = "Reprovado(a)";
+                                $situation = "Reprovado";
                             } else {
                                 $situation = "Recuperação";
                             }
                         } else {
-                            $situation = "Reprovado(a)";
+                            $situation = "Reprovado";
                         }
                     }
                 }
@@ -620,9 +620,35 @@ class EnrollmentController extends Controller
                         }
                     }
                 }
-                $gradeResult->situation = "Aprovado(a)";
+                $gradeResult->situation = "Aprovado";
                 $gradeResult->save();
             }
+            //Mudar status da matrícula
+            //1 = Em andamento; 6 = Aprovado; 8 = Reprovado
+            if ($studentEnrollment->status == "1" || $studentEnrollment->status == "6" || $studentEnrollment->status == "8") {
+                $allGradesFilled = true;
+                $situation = "Aprovado";
+                foreach ($studentEnrollment->gradeResults as $gradeResult) {
+                    if ($gradeResult->situation == null) {
+                        $allGradesFilled = false;
+                    } else if ($gradeResult->situation == "Reprovado") {
+                        $situation = "Reprovado";
+                        break;
+                    } else if ($gradeResult->situation == "Recuperação") {
+                        $situation = "Em Andamento";
+                        break;
+                    }
+                }
+                if ($allGradesFilled && $situation == "Aprovado") {
+                    $studentEnrollment->status = 6;
+                } else if ($situation == "Reprovado") {
+                    $studentEnrollment->status = 8;
+                } else {
+                    $studentEnrollment->status = 1;
+                }
+                $studentEnrollment->save();
+            }
+
         }
     }