+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 ## [Versão 3.37.75]
 - Adicionado o campo pcd no relatório de matrícula
 
-## [Versão 3.36.75]
+## [Versão 3.37.76]
+- Relatório Dados Estatísticos
+
+## [Versão 3.36.76]
 - O botão inoperante de calcular média na página de notas foi removido
 
 ## [Versão 3.36.75]